requirements.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.4
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.14
